Fixed ID register address
This was an oversight. The last address in the virtual Wii memory space
(0xFE) does not match the address register ON the device, which is 4
bytes back. So I was getting the 16-bit IDs I expected, but they were at
the END of the block rather than the front. I chalked it up to a quirk,
turns out I was just reading the wrong position.