Updated whammy bar range comment
So this is interesting. On my GH3 guitar, the whammy bar starts
physically at "center" and can only be moved towards the body of the
guitar. But the potentiometer reading starts at center in the range, so
presumably one could remove the stop and use the bar in both directions?
Or perhaps the World Tour guitar uses the full range? I'm leaving the
return value unadjusted just in-case.