Revert guitar ID fix (d30f0f2)
While I was right about the ID byte being '0', this was only true if the
guitar was initialized, polled for an update, and then re-initialized
without being disconnected. This would happen often during testing, but
is unlikely to happen in general use. Better to check for the actual ID
(0x01), and the user can disable enforcement if necessary.