Disable enforceID for shared data instances
If you're sharing data, the communication functions (begin, connect,
update, etc.) will return errors if the ID doesn't match, when what
you're concerned about is a communication error. It only makes sense to
let the user handle incorrect IDs and report comm errors only. The user
can always switch this on for each class if need-be.