Enable enforceID for shared data instances
Reverts 67424ed. While the logic for the previous commit is sound,
silently disabling the ID enforcement makes the function calls behave
differently, which would be confusing to the end-user. The user can
disable this setting if need-be, and otherwise run the comms-only calls
through the generic parent class if sharing data.