Replaced Stream references with Print
For Arduino, Stream inherits from Print. The *print* class is what I actually want to be printing to, and since Serial is part of Print this works out. Plus it means that any community libraries (devices etc.) that use the Print class can be printed-to using the library's debug functions.