Moved identifyController to private
This is a redundant function to the end user. Because the controller needs to be initialized (read: connect()-ed) before it will report a proper ID, the user should just use getControllerType and avoid the redundant bus update. Still required for the class internally to fetch the type from the controller during connect().