Remove 'reconnect' function
This was more confusing than anything, since it's not clear to the user what the difference between "connect" and "reconnect" is. The "reconnect" function just didn't clear out the control data and connected id. But then again, why would you ever want to preserve the old control data if you're trying to reconnect to a controller?

Having just the one function is simpler and more straight-forward.