Make inline comms functions protected
In hindsight I think this was a little misguided... By having these member functions publicly accessible outside of the normal control flow, the shared data is not updated as you would expect (notably with 'identifyController'). The user might also expect these to use the default I2C port instead of the member for the object, which is not necessarily intuitive.

Instead, I'm making these protected so they are still available to children (see 'specialInit') and explicitly "for internal use only".