Modify ClassicController ID check for double init
This should ("should") never happen during normal use. From testing the 0x00 report for byte 4 happens when a read request is interrupted and the controller is reinitialized without power cycling it. In other words, this happens during development when the microcontroller is flashed with the controller still connected. But during normal use the controller will always power cycle before it's reinitialized.

This doesn't seem to affect anything else, so I'm adding it in to account for that edge case.