Add "High Resolution" mode for ClassicController
All inputs now flip between "standard" and "high resolution" based on the saved setting. Calling the "setHighRes" function can enable or disable this mode by writing a register value on the controller.

This is "cleaner" on the frontend compared to the template version (one class instead of two) and gives you the flexibility to switch between "modes" at will, although the resulting class now contains twice as many maps...