Move requestSize to data struct
The request size should be shared by all controller objects using the same logical "port". Only one controller can be active at a time with any given shared data group, so it doesn't make sense to keep this exclusive to the instance.

This also means that when multiple controller types are used with shared data, changes in the request size  will also transfer to the generic ExtensionPort object. Important for specificInit usage when using multiple types.