Reset request size on connection
Finally deciding to pull the trigger on this. Previously, requestSize was completely up to the user and wasn't managed by the library at all because all controllers support the 6-byte request mode.

The existence of the Classic Controller "high res" mode creates a conflict, because connecting to the controller in that mode requires an 8 byte request size which then persists after disconnecting even for controllers that don't make use of it.

So I'm changing the idiom. The *library* now manages request size if it needs to. If the user wants to change the request size, they can do that after connection. This makes the library more efficient and flexible. Plus even user-set request sizes should be different for each controller type, which would be set after the controller is initialized (read: identified) anyways.