Make I2C member functions public
This allows users to write arbitrary data to the controller outside of the controller-specific class. Using this functions in the wrong place has the possibility to "break" the controller data flow. User beware...

The advantage of course is that users can write and read arbitrary data without having to build a custom class for minor tweaks.