Use pointer argument for checkHighRes function
The function now returns the success state of the check: true if it successfully read the 'mode' of the controller (std or high res) and 'false' if there was an I2C communication error. The 'mode' state of the controller is instead set to the bool argument passed by pointer.

This prevents accidentally setting the internal 'high res' mode to 'false' if there was a communication error, rather than the function reading the 'standard' map.