Refactor internal "highRes" funcs to "dataMode"
This isn't explicitly about high resolution mode anymore, but about reading the current data mode of the controller.

The public-facing functions will still be kept as "highRes" because it's more idiomatic to check whether a specific mode is 'true' rather than trying to match up the boolean values to a specific mode.