Classic: Re-attempt connect if HR write fails
Some Classic clone controllers apparently NACK the register write for "high res" mode which wasn't known when they were first reverse engineered. The result is that the specificInit returns 'false' indicating a communication error and the controller fails to indicate that it's connected (#68).

This commit should hopefully fix that issue by attempting to distinguish between controllers that NACK the register write for "high res" mode because they don't support it, and controllers that NACK the register write because they're no longer connected to the bus. This is done with a simple read of the control data. If we receive data the controller is still present. If not, the mode is unsupported.

Note that this can't distinguish between unsupported data and momentary (sub ms) loss of connection. That's just something we're going to have to live with for the time being...