Fix BuildControllerClass initialization order
According to the standard, data members are initialized after base classes which means this shouldn't work properly (depending on what's included in the data struct and how it's accessed).

To fix the issue, we inherit from the data struct as private and then pass the (initialized!) version of the struct (static cast 'this' pointer) to the ExtensionController object, which we also inherit.